dist[mm]をdouble型からint型へ変更. oiledge-debug を oiledge-sample にリネーム.

+++ src/oilEdgePosSsm.h
@@ -4,7 +4,7 @@
 #define SNAME_OILEDGEPOS "oilEdgePos"
 
 typedef struct{
-	double dist;
+	int dist;
 	double theta;
 	int status;
 }OilEdgePos;
